Switch to (modified) hugo-xmin theme
This change comes also with a switch from a .toml config to a .yaml one.
From what I can tell, hugo allows the formats to be used
interchangeably, but I was too lazy to transfer the hugo-xmin example
config from yaml to toml.

Also, I took the liberty of adding empty lines between markdown headings
to be more conforming to spec. And I added tables of content for every
file with headings. hugo-xmin supports them with a few tweaks that are
already included in my fork of the theme's repo.